Stb 4

+++ sport-telegram-bot.Application/Features/TrainRecord/Queries/GetActiveTrainsByUser/GetActiveTrainsByUserRequest.cs
@@ -3,5 +3,5 @@ using MediatR;
 
 namespace sport_telegram_bot.Application.Features.TrainRecord.Queries.GetActiveTrainsByUser
 {
-    public sealed record GetActiveTrainsByUserRequest(int UserId) : IRequest<IReadOnlyList<Domain.TrainRecord>>;
+    public sealed record GetActiveTrainsByUserRequest(long UserId) : IRequest<IReadOnlyList<Domain.TrainRecord>>;
 }
